fix: split v1 hardfork tests to different files
